requirements.txt
Dependencies (versions are my own):
-----------------------------------

einops: 0.8.0
matplotlib: 3.7.2
mdtraj: 1.10.0
networkx: 3.1
numpy: 1.24.3
scikit-learn: 1.3.2
scipy: 1.10.1
tensorboard: 2.12.1
pytorch: 2.4.1
pytorch-cuda: 11.8
pytorch-mutex: 1.0
torchdiffeq: 0.2.2
torchvision: 0.19.1
tqdm: 4.66.5

